Fibonacci Modified - small adjustment in the test of it.

+++ src/test/java/com/trainings/algorithms/dynamicprogramming/FibonacciModifiedTest.java
@@ -2,11 +2,13 @@ package com.trainings.algorithms.dynamicprogramming;
 
 import org.junit.Test;
 
+import java.math.BigInteger;
+
 import static org.junit.Assert.*;
 
 public class FibonacciModifiedTest {
     @Test
     public void example1(){
-        assertEquals(27, FibonacciModified.fibonacciModified(0,1,6));
+        assertEquals(BigInteger.valueOf(27), FibonacciModified.fibonacciModified(0,1,6));
     }
 }
